fix(duration): widen Duration.days from short to int — 89y silent-truncation bug
Duration.getDays() / DTD.days / Dur.days were 16-bit signed shorts. Any
xs:dateTime subtraction across more than Short.MAX_VALUE = 32 767 days
(~89.7 years) silently truncated mod 65 536, producing arithmetically
wrong durations. Bitemporal queries on long-lived histories (medical
records, land registries, archival document stores) hit this in
production despite being entirely valid xs:dateTime arithmetic.

Widens Duration.days to int (Integer.MAX_VALUE ≈ 5.88M years) and
removes the parser-time Short.MAX_VALUE checks. Constructor and parser
signatures shift from short to int for the days field; getDays() returns
int now. The high-bit-encoded sign on months is unchanged.

AbstractTimeInstant.subtract still defends against a long->int narrowing
overflow with ERR_OVERFLOW_UNDERFLOW_IN_DURATION, but that branch is
unreachable from any pure-Java entry point because LocalDate caps year
at 999_999_999.

Tests:
* property_dateTimeSubtract_preservesSignedInstantDifference exercises
  ±Short.MAX_VALUE × 4 days (~±359 years) over 10 000 random pairs.
  Pre-widening these would have wrapped silently.
* property_dtdAdd / property_dtdSubtract / property_ymdAdd /
  property_ymdSubtract / property_dtdAdd_resultIsCanonical at 10 000
  iters each prove the addInternal total-micros and total-months
  invariants from docs/formal-verification.md (Inv 1.1a, 1.2a, 1.3a,
  1.5a) hold across the widened range.
* All 1042 Brackit tests pass.

The randomized DateTime subtract test uses a Supplier<String> message
form for assertEquals — the eager String form was constructing 10 000 ×
3 toString() calls per iteration, generating enough GC pressure to make
the test appear hung at 90 s+ even though each iteration is O(1).

+++ src/main/java/io/brackit/query/atomic/AbstractDuration.java
@@ -70,7 +70,7 @@ public abstract class AbstractDuration extends AbstractAtomic implements Duratio
     StringBuilder out = !isNegative() ? new StringBuilder("P") : new StringBuilder("-P");
     short years = getYears();
     byte months = getMonths();
-    short days = getDays();
+    int days = getDays();
     byte hours = getHours();
     byte minutes = getMinutes();
     int micros = getMicros();
